Move and rename virtualEnvironmentIdentifier.ts
