+++ src/client/pythonEnvironments/common/environmentManagers/poetry.ts
@@ -8,7 +8,7 @@ import { traceError, traceVerbose } from '../../../common/logger';
 import { getOSType, getUserHomeDir, OSType } from '../../../common/utils/platform';
 import { getPythonSetting, isParentPath, pathExistsSync, readFileSync, shellExecute } from '../externalDependencies';
 import { getEnvironmentDirFromPath } from '../commonUtils';
-import { isVirtualenvEnvironment } from '../../discovery/locators/services/virtualEnvironmentIdentifier';
+import { isVirtualenvEnvironment } from './simplevirtualenvs';
 import { StopWatch } from '../../../common/utils/stopWatch';
 import { cache } from '../../../common/utils/decorators';
 import { isTestExecution } from '../../../common/constants';
