+++ src/test/pythonEnvironments/discovery/locators/virtualEnvironmentIdentifier.unit.test.ts
@@ -14,7 +14,7 @@ import {
     isVenvEnvironment,
     isVirtualenvEnvironment,
     isVirtualenvwrapperEnvironment,
-} from '../../../../client/pythonEnvironments/discovery/locators/services/virtualEnvironmentIdentifier';
+} from '../../../../client/pythonEnvironments/common/environmentManagers/simplevirtualenvs';
 import { TEST_DATA_ROOT, TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
 import { assertVersionsEqual } from './envTestUtils';
